backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yuelin-official-site-api"
version = "0.1.0"
description = "FastAPI backend for the Shenzhen Yuelin Technology official site."
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16,<2.0",
  "email-validator>=2.2,<3.0",
  "fastapi>=0.116,<1.0",
  "pydantic-settings>=2.10,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.4,<9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
